+++ src/axpel/components/Map.tsx
@@ -1,9 +1,10 @@
 import React, { ReactElement } from 'react'
 import { GoogleMap, Marker } from '@react-google-maps/api';
+import { useSelector, shallowEqual,} from "react-redux";
 
 const containerStyle = {
   width: '100%',
-  height: '70vh'
+  height: '60vh'
 };
 
 const center = {
@@ -11,17 +12,12 @@ const center = {
   lng: 0.0
 };
 
-const position = {
-    lat: 37.772,
-    lng: -122.214
-  }
-
-  
-  const onLoad = (marker: any) => {
-    console.log('marker: ', marker)
-  }
 
 function MyComponent() : ReactElement {
+
+  const savedPositions = useSelector((state: any) => state.maps.savedPlaces, shallowEqual)
+    
+
   return (
 
       <GoogleMap
@@ -29,11 +25,17 @@ function MyComponent() : ReactElement {
         center={center}
         zoom={2}
       >
-        { /* Child components, such as markers, info windows, etc. */ }
-        <>
-        <Marker  onLoad={onLoad}
-        position={position} />
-        </>
+        {savedPositions?.map( (savedPosition :any , index :any) => 
+                  <Marker
+                  key={index}
+                  position={
+                    {
+                      lat: savedPosition.geometry.location.lat(),
+                      lng: savedPosition.geometry.location.lng()
+                    }
+                  } />
+        )}
+
       </GoogleMap>
   )
 }