+++ src/axpel/screens/Exercise-2.tsx
@@ -1,28 +1,37 @@
-import React, { ReactElement } from 'react'
+import React, { Profiler, ReactElement } from 'react'
 import Footer from '../../components/elements/Footer/Footer'
 import Header from '../../components/elements/Header/Header'
 import Map from '../../axpel/components/Map'
 import Searchbar from '../components/Searchbar'
 import { LoadScript } from '@react-google-maps/api'
 
-import { useSelector, shallowEqual } from "react-redux";
 
 export default function Exercise2(): ReactElement {
 
-    const reduxTest = useSelector((state: any) => state.maps, shallowEqual)
-
-    console.log("Exercise2", reduxTest)
+    // https://reactjs.org/docs/profiler.html
+   let onRenderCallback : any  = (
+        id: any, // the "id" prop of the Profiler tree that has just committed
+        phase: any, // either "mount" (if the tree just mounted) or "update" (if it re-rendered)
+        actualDuration: any, // time spent rendering the committed update
+        baseDuration: any, // estimated time to render the entire subtree without memoization
+        startTime: any, // when React began rendering this update
+        commitTime: any, // when React committed this update
+        interactions: any // the Set of interactions belonging to this update
+      ) => {
+        // Aggregate or log render timings...
+        console.log( 'id -', id, ' The phase - ', phase, 'Time taken for the update', actualDuration, baseDuration, startTime, commitTime, interactions);
+      }
 
     return (
         <>
-            <Header />
+        <Profiler id="exercise2" onRender={onRenderCallback} >
+            <Header title="Axpel Search App" link="https://github.com/martinrebo/boilerplate/tree/main/src/axpel" linkText="README" />
             <LoadScript googleMapsApiKey="AIzaSyBxwQlDE8omqzejYDuo_ZmycEh8QQjM5GY" libraries={["places"]}>
                 <Searchbar />
                 <Map />
             </LoadScript>
-
             <Footer title="Axpel Map" subtitle="Challenge 2: Google Map Suggestions" />
-
+            </Profiler>
         </>
     )
 }